BjlxBS/requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
altgraph==0.17.3
anyio==3.6.2
appdirs==1.4.4
Appium-Python-Client==2.8.1
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
bcrypt==4.0.1
bottle==0.12.23
cachetools==5.2.0
cchardet==2.1.7
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
constantly==15.1.0
cryptography==38.0.1
customtkinter==5.2.0
darkdetect==0.8.0
ddt==1.6.0
docopt==0.6.2
et-xmlfile==1.1.0
exceptiongroup==1.0.0rc9
fastapi==0.85.1
Flask==2.2.2
future==0.18.2
h11==0.14.0
hpack==4.0.0
html-testRunner==1.2.1
hyperframe==6.0.1
idna==3.4
importlib-metadata==5.0.0
incremental==22.10.0
iniconfig==1.1.1
ipython-genutils==0.2.0
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
jsonpath==0.82
MarkupSafe==2.1.1
mistune==2.0.4
mpmath==1.2.1
multidict==6.0.2
network==0.1
numpy==1.23.4
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandocfilters==1.5.0
parse==1.19.0
parso==0.8.3
pathtools==0.1.2
pbr==5.11.0
pefile==2022.5.30
pickleshare==0.7.5
pip-review==1.2.0
pipreqs==0.4.13
playsound==1.3.0
pluggy==1.0.0
priority==2.0.0
prometheus-client==0.15.0
psutil==5.9.3
py2exe==0.12.0.1
pyasn1==0.4.8
PyAutoIt==0.6.5
pycparser==2.21
pydantic==1.10.2
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
PyMySQL==1.0.2
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PyQt5Designer==5.14.1
pyserial==3.5
PySocks==1.7.1
pytest==7.2.0
pythonping==1.1.4
pytz==2022.5
pywin32==304
pywin32-ctypes==0.2.2
pywinauto==0.6.8
pywinpty==2.0.8
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
ruamel.appconfig==0.5.5
ruamel.std.argparse==0.8.3
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
schedule==1.1.0
scipy==1.9.3
selenium==4.5.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
starlette==0.20.4
sympy==1.11.1
testpath==0.6.0
toml==0.10.2
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
win32-setctime==1.1.0
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
xmltodict==0.13.0
yaml-1.3==0.1.0
yarg==0.1.9
yarl==1.8.1
zipp==3.10.0
